Initial implementation of AWX healing action
This patch completes the initial implementation of healing using an AWX
(a.k.a. Ansible AWX, or Ansible Tower) job. The mechanism is based in
the definition of a _healing rule_ like this:

  apiVersion: monitoring.openshift.io/v1alpha1
  kind: HealingRule
  metadata:
    name: myhealingrule
  spec:
    conditions:
    - alert: myalert
    actions:
    - awxJob:
        address: https://myawx.example.com/api
        secret: mysecret
        project: "My Project"
        template: "My Job Template"

The `alert-healer` component will watch the collection of `Alert`
objects. When a new alert is added or modified it will compare it with
all the existing healing rules, and will execute the actions of those
that match. For example, the above healing rule will look for alerts
whose name is `myalert`, when there is one that matches it will connect
to the AWX server in `myawx.example.com` and it will create a new job
from the template `My Job Template` inside the project `My Project`.

The credentials to connect to the AWX API will be obtained from the
secret `mysecret`, which should contain the `username` and `password`
entries. For example:

  apiVersion: v1
  kind: Secret
  metadata:
    name: mysecret
  data:
    username: YWRtaW4=
    password: bXlwYXNzd29yZA==

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>