Simplify UIDs
Currently the images use a complicated workaround to be able to assign
an stable numeric user ID to the `monitoring` user. The unusually large
UID causes problems when building en certain environments, in particular
in Travis CI. To avoid that this patch changes the images so that they
don't specify the user.

Note that this simplifies the development environment, but for
production environments, when using external storage, it will be
probably necessary to put in place a better mechanism to ensure that the
user that runs the binaries can write to the storage.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>