Make alert names unique
Currently the name of published alerts is calculated from the name of
the rule. This means that two alerts that have different labels
(different nodes going down, for example) will result in the same
published alert. This makes it difficult to apply healing actions that
are specific to each alert (starting only the node that went down, for
example). To improve that this patch changes the alert publisher so that
it will calculate the name from the alert rule and from a hash of the
attributes (labels and annotations) that identify it.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>