Improve templating of actions
Currently the definition of actions are processed as templates. To do so
we convert them to a JSON document, then we process that JSON document
as template and finally we convert that JSON document back to an action
object. This introduces problematic cases because the JSON quoting
interferes with the templating language. In addition it is difficult to
process Ansible playbooks as templates because they use the same
delimiters than Go templates. To address those two problems this patch
changes the approach; instead of converting the to JSON the object is
now iterated using reflection, and each string is processed
independently.

The delimiters can be now explicitly specified in the definition of the
action. For example, for an Ansible playbook it is very convenient to
not use {{ and }} as delimiter. To use [[ and ]] instead the action can
be defined as follows:

  actions:
  - delimiters:
      left: "[["
      right: "]]"
    ansiblePlaybook: |-
      ---
      - name: Ensure oVirt node
        hosts: localhost
        connection: local
        gather_facts: false
        pre_tasks:
        tasks:
        - name: Ensure node [[ $labels.instance ]]
          ovirt_vms:
            name: {{ [[ $labels.instance ]] | regexp_replace('\..*$', '') }}
            state: running

Signed-off-by: Juan Hernandez <jhernand@redhat.com>