Fix creation of Ansible playbook jobs
Currently the jobs created to run Ansible jobs use the name of the alert
as the name of the job.  This doesn't work because names of alerts
usually contain mixed case, like in "NodeDown", and the name of the job
has to be a valid Kubernetes object name, where only lower case is
accepted. To address that this patch changes the way that job names are
generated: they will be calculated from the name of the alert converted
to lower case, plus a suffix containing a hash of the alert data, to
make it unique.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>