Make jobs owners of configuration maps
Currently the configuration maps that are created to contain the
playbooks and the inventory for Ansible jobs aren't owned by the jobs.
This means that when the job is deleted the configuration map isn't
automatically deleted. To improve that this patch changes the healer so
that once the job is created the configuration map will be updated to
set it as owned by the job.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>