Remove the Prometheus launcher
Currently this project contains the code for the auto-heal service and
also for the `prometheus-laucher`. This launcher is responsible for
reading definitions of alerting, building the Prometheus configuration,
and reloading it. That has been replaced by the Prometheus operator, so
it no longer makes sense to have it in this repository. This patch
removes it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>